+++ app/src/main/AndroidManifest.xml
@@ -25,7 +25,7 @@
         </activity>
 
         <activity
-            android:name=".RvCatering"
+            android:name=".RV.RvCatering"
             android:exported="false">
             <meta-data
                 android:name="android.app.lib_name"
@@ -34,7 +34,7 @@
 
 
         <activity
-            android:name=".RvStreetFood"
+            android:name=".RV.RvStreetFood"
             android:exported="false">
             <meta-data
                 android:name="android.app.lib_name"
@@ -42,7 +42,7 @@
         </activity>
 
         <activity
-            android:name=".RvRestaurants"
+            android:name=".RV.RvRestaurants"
             android:exported="false">
             <meta-data
                 android:name="android.app.lib_name"
