+++ app/src/main/java/com/example/forfoodiesbyfoodies/RV/RvRestaurants.java
@@ -1,15 +1,18 @@
-package com.example.forfoodiesbyfoodies;
-
-import android.annotation.SuppressLint;
-import android.content.Intent;
-import android.os.Bundle;
+package com.example.forfoodiesbyfoodies.RV;
 
 import androidx.annotation.NonNull;
 import androidx.appcompat.app.AppCompatActivity;
 import androidx.recyclerview.widget.LinearLayoutManager;
 import androidx.recyclerview.widget.RecyclerView;
 
+import android.content.Intent;
+import android.os.Bundle;
+
+import com.example.forfoodiesbyfoodies.DetailsView;
 import com.example.forfoodiesbyfoodies.Entities.FoodPlace;
+import com.example.forfoodiesbyfoodies.MyAdapter;
+import com.example.forfoodiesbyfoodies.R;
+import com.example.forfoodiesbyfoodies.RecycleViewInterface;
 import com.google.firebase.database.DataSnapshot;
 import com.google.firebase.database.DatabaseError;
 import com.google.firebase.database.DatabaseReference;
@@ -18,9 +21,8 @@ import com.google.firebase.database.ValueEventListener;
 
 import java.util.ArrayList;
 
-public class RvCatering extends AppCompatActivity implements RecycleViewInterface {
-
-
+public class RvRestaurants extends AppCompatActivity implements RecycleViewInterface
+{
 
 
     DatabaseReference dataPlaces;
@@ -28,18 +30,19 @@ public class RvCatering extends AppCompatActivity implements RecycleViewInterfac
     ArrayList<FoodPlace> list;
     DatabaseReference databaseReference;
     MyAdapter adapter;
-    String place= "Catering";
+    String place= "Restaurant";
 
-    @SuppressLint("MissingInflatedId")
+
     @Override
     protected void onCreate(Bundle savedInstanceState) {
         super.onCreate(savedInstanceState);
-        setContentView(R.layout.rv_catering);
+        setContentView(R.layout.rv_restaurants);
 
 
         dataPlaces = FirebaseDatabase.getInstance().getReference();
-        recyclerView=findViewById(R.id.recycle_catering);
+        recyclerView=findViewById(R.id.recycle_restaurants);
         databaseReference= FirebaseDatabase.getInstance().getReference(place);
+
         list = new ArrayList<>();
         recyclerView.setLayoutManager(new LinearLayoutManager(this));
         adapter = new MyAdapter(this, list,this);
@@ -63,12 +66,11 @@ public class RvCatering extends AppCompatActivity implements RecycleViewInterfac
         });
 
 
-
     }
 
     @Override
     public void onItemClick(int position) {
-        Intent intent = new Intent(RvCatering.this, DetailsView.class);
+        Intent intent = new Intent(RvRestaurants.this, DetailsView.class);
         startActivity(intent);
 
     }
